New Objectmapper for ApiDto and output Dto

+++ src/main/java/com/erlis/weather/dto/api/TemperatureDto.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 
 @Data
 public class TemperatureDto implements Serializable {
-    public Integer temp;
+    public Double temp;
     public Double temp_min;
     public Double temp_max;
     public Integer pressure;
